requirements.txt
configparser==5.0.0
python-telegram-bot==12.7
psycopg2==2.8.5